+++ src/utils/DateUtils.ts
@@ -34,3 +34,11 @@ export const dateToValue = (date: Dayjs, endOfDay: boolean = false): number => {
 
 // Utility function to convert a numeric value back to a date
 export const valueToDate = (value: number): Dayjs => dayjs(value);
+
+/** Calendar day → YYYYMMDD integer (local calendar fields from dayjs). */
+export const dayjsToDayPeriod = (d: Dayjs): number =>
+  d.year() * 10000 + (d.month() + 1) * 100 + d.date();
+
+/** Calendar month → YYYYMM integer. */
+export const dayjsToMonthPeriod = (d: Dayjs): number =>
+  d.year() * 100 + (d.month() + 1);
